Removed late-socials.js's reference to $("#share-container"); it's now passed in, so that the code is completely decoupled from index.html.
Used $.each() to iterate over the keys in socialMedia, avoiding the need for another non-inline function (and its associated comments).

Also removed accidental 'medium' global.

Changed the constructor signature for SocialMedia() so that a function called getURL() is passed in, which returns the URL to share (the original purpose of 'selector' was confusing me so I thought getURL() was more straightforward).