Added a hack to fix #22, "Test suite doesn't run on IE9".
I'm not a fan of this hack, or the inclusion of the base href tag
in the first place (which was added in
d5d3ce3f6c7ad470e9a2051af5ae852aa6bdab15). It'd be nice if we could
have that codemirror module wrapper not depend on the base href of the
page it's loaded into, but this will do for now.