Pragmatic pushState() fix for making role of address bar URL clearer.
This is not the same as #30, which aims to add full support for
pushState(). Instead, here we implement "half" of that, using pushState()
when the user publishes their work--to ensure that the page is what
they expect it to be if they share/bookmark their URL. HOWEVER, if the
user clicks the back button, we just issue a full-page reload, since the
rest of the code isn't built for unexpected URL changes yet.

Because it relies on pushState(), this doesn't work on IE9 yet.