Simplified relocator by using more CodeMirror methods.
Specifically, used CodeMirror's addWidget(), setLineClass(), and
setMarker() to obviate the need for lots of custom DOM/style
manipulation that will potentially break when we upgrade
CodeMirror.

Because we're now using setLineClass(), this also resolves
we set on lines.

There are still a few wonky things going on with this commit, such
as the fact that information always appears below the line that
help displays on, and sometimes this results in weird scrolling
jumps, but we can resolve this in later commits.