Point error/help messages at the correct soft-wrapped line.
Because we're using soft-wrapped lines, it can be the case that an error
or help message is pointing at a line that is several physical screen
lines down from where the line of code actually starts. Until now,
the messages always pointed at the first physical screen line, which
meant that it sometimes obscured the actual character it was supposed
to be pointing at.

I also made the error/help messages overlap the gutter, in part because
this will make the errors display better on narrow screens, and in part
because this will allow us to actually have the arrow point at the
relevant character in question.