Move CSS out of fc/ui/text.js. "normal" now default text size.
Moving the CSS out of fc/ui/text.js also simplifies it a lot, as setting
text size is now just a matter of toggling some classes on the CodeMirror
wrapper, rather than dynamically generating a stylesheet.

This fixes #86.

Also added some code to deal with edge cases if localStorage is full.