Much better indication for error targets ("here" labels).
Specifically, the first "here" is not highlighted at all because
the error message is pointing at it. The second "here" appears
hyperlinked, and if the user clicks on it, the cursor position
is set to the location, and an animated red arrow appears to
guide the user's eye towards it.