If text widget throws an exception on first-run, test suite will now catch it.
The initialization code is now more "crash early, crash often" in a way
that will be caught by the testing suite if anything is amiss. For instance,
if 'data-default-size' isn't on one of the text sizes to specify a
default text size for the app, the test suite will now fail.