Moved CSS out of js/fc/ui/text.js and into CSS files.
The TextUI code is now much smaller, and it's also easier to understand what the
CSS for the various text size settings are.

Note that this is just a pure refactoring--I've tried to create the static
CSS in a way that's as faithful to the dynamically-generated CSS as possible.
Glancing at the whole CSS file, it seems that some rules are now redundant,
but we can save those fixes for a later commit.