Allow Typekit to be disabled from querystring for testing purposes.
Right now Typekit appears to take a really long time to fail if
it's loaded on either an unsupported browser (like Opera, apparently)
or from a domain that isn't authorized. While we should generally fix
this to be more responsive for users on unsupported browsers, for now
it's making some unit/integration tests take a *lot* longer to run
than they ought to, so I'm committing this as a stopgap measure.

This is also probably related to #103, "Load Typekit asynchronously".