The confirm/publish dialogs now always appear under the publish button.
In order to do this, I changed the animation from a slide to a fade
because the addition of margins to the .thimble-modal-menu selector
made the slide appear wonky.

We can fix this later; Jess also seemed to want to change the publish UI
a lot, so I didn't want to spend too much time figuring out how to
retain the slide at its original quality. Also, a fade transition still
looks okay.