Factored out EditorWithToolbar class.
This just combines the editor with the toolbar and changes the toolbar
class to not display the Publish button if no startPublish callback
has been set. This refactoring provides a better separation of concerns
between editing and publishing, and allows the EditorWithToolbar widget
to be embedded in a webpage if no publish functionality is needed.