Fixed test suite's spurious 'a.contentWindow is null' exception.
This is raised as 'Cannot read property "document" of null' in WebKit.

This bug occurred because some tests would very	briefly
instantiate a two-paned	editor with live-preview and then destroy it
by removing the preview from the DOM. The parsing-codemirror triggers a
reparse on delay, and unfortunately the delay would be called after
the preview was	removed	from the DOM, which would trigger this error.

Instead of raising such an inscrutable exception, LivePreview now
detects if the preview area isn't attached to the document, and if so,
logs a console message and returns.